requirements.txt
flask
flask-cors
numpy
Pillow
face-recognition
ultralytics
supabase
python-dotenv
